# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "USER/TO-DO CRUP OPERATION",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer' followed by a space and then your token. Example: 'Bearer abc123'"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/toDo/data": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "Created",
            "examples": {
              "application/json": {
                "message": "here is your all todo items",
                "todo": {
                  "email": "update@gmail.com",
                  "contact": "4567890123",
                  "title": " title",
                  "discription": "discription"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json": {
                "message":" email is require ,enter you updated email "
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      }
    }
  },
  "/toDo/upd": {
    "patch": {
      "description": "",
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "example": "update@gmail.com"
              },
              "contact": {
                "example": "2345678901"
              },
              "title": {
                "example": "updated title"
              },
              "discription": {
                "example": "updated discription"
              }
            }
          }
        }
      ],
      "responses": {
        "200": {
          "description": "OK",
          "examples": {
            "application/json": {
              "message": "updated",
              "todoUpdate": {
                "email": "update@gmail.com",
                "contact": "updated contact",
                "title": "  updated title",
                "discription": " updated discription"
              }
            }
          }
        },
        "403": {
          "description": "Bad Request",
          "examples": {
            "application/json": {
              "message1":" email field should not be empty ,ennter valid email ",
              "message2":"User not authorized to update this todo."
          }
        }
        },
        "404": {
          "description": "Not Found",
          "examples": {
            "application/json": {
              "message":"atleast one field is requuire to update details"
          }
        }
        },
        "500": {
          "description": "Internal Server Error",
          "examples": {
            "application/json": {
              "message": "internal servar error"
          }
        }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "/toDo/del": {
    "delete": {
      "description": "",
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "example": "update@gmail.com"
              }
            }
          }
        }
      ],
      "responses": {
        "200": {
          "description": "OK",
          "examples": {
            "application/json": {
              "message":"deleted" 
          }
        }
        },
        "404": {
          "description": "Not Found",
          "examples": {
            "application/json": {
              "message": "enter updated email"  
          }
        }
        },
        "500": {
          "description": "Internal Server Error",
          "examples": {
            "application/json": {
              "message":"internal server error,try to login" 
          }
        }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  }
}
}